+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsContent.tsx
@@ -1,38 +1,33 @@
 import { Divider, Layout } from 'antd'
 
 import {
-  V2SettingsContentKey,
+  V2SettingsKey,
   V2SettingsKeyTitleMap,
-} from 'components/v2/V2Project/V2ProjectSettingsPage/V2ProjectSettings'
+} from 'components/v2/V2Project/V2ProjectSettings/V2ProjectSettings'
 
-import V2ProjectSettingsVenftContent from 'components/v2/V2Project/V2ProjectSettingsPage/V2ProjectSettingsVenftContent'
-
+import V2ProjectSettingsVenftContent from 'components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsVenftContent'
 import ProjectPayersSection from 'components/Project/ProjectToolsDrawer/ProjectPayersSection'
-
 import { TransferOwnershipForm } from 'components/Project/ProjectToolsDrawer/TransferOwnershipForm'
-
-import V2ReconfigureProjectHandle from 'components/v2/V2Project/V2ProjectSettingsPage/V2ReconfigureProjectHandle'
-
-import V2ProjectDetails from 'components/v2/V2Project/V2ProjectSettingsPage/V2ProjectDetails'
-
+import V2ReconfigureProjectHandle from 'components/v2/V2Project/V2ProjectSettings/V2ReconfigureProjectHandle'
+import V2ProjectDetails from 'components/v2/V2Project/V2ProjectSettings/V2ProjectDetails'
 import { EditTokenAllocationContent } from 'components/v2/V2Project/EditTokenAllocationContent'
-import V2ProjectSettingsPayoutsContent from 'components/v2/V2Project/V2ProjectSettingsPage/V2ProjectSettingsEditPayouts'
+import V2ProjectSettingsPayoutsContent from 'components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsEditPayouts'
 import ArchiveV2Project from 'components/v2/V2Project/ArchiveV2Project'
 import { V1TokenMigrationSetupSection } from 'components/v2/V2Project/V2ProjectToolsDrawer/V1TokenMigrationSetupSection'
 import { useRouter } from 'next/router'
 import { useState, useEffect } from 'react'
-import V2ProjectReconfigure from 'components/v2/V2Project/V2ProjectSettingsPage/V2ProjectReconfigure'
+import V2ProjectReconfigure from 'components/v2/V2Project/V2ProjectSettings/V2ProjectReconfigure'
 
-const defaultPage: V2SettingsContentKey = 'general'
+const defaultPage: V2SettingsKey = 'general'
 
 const V2ProjectSettingsContent = () => {
   const [selectedSettingsPage, setSelectedSettingsPage] =
-    useState<V2SettingsContentKey>(defaultPage)
+    useState<V2SettingsKey>(defaultPage)
   const router = useRouter()
 
   useEffect(() => {
     setSelectedSettingsPage(() => {
-      switch (router.query.settingsPage) {
+      switch (router.query.page) {
         case 'general':
           return 'general'
         case 'projecthandle':
@@ -57,9 +52,9 @@ const V2ProjectSettingsContent = () => {
           return defaultPage
       }
     })
-  }, [router.query.settingsPage])
+  }, [router.query.page])
 
-  const getActiveTab = (selectedSettingsPage: V2SettingsContentKey) => {
+  const getActiveTab = (selectedSettingsPage: V2SettingsKey) => {
     switch (selectedSettingsPage) {
       case 'general':
         return <V2ProjectDetails />
