+++ src/components/v2/V2Project/index.tsx
@@ -3,15 +3,13 @@ import ProjectHeader from 'components/Project/ProjectHeader'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 // TODO: Do we still need lazy loading?
 
-import { useContext, useEffect, useState } from 'react'
+import { useContext, useState } from 'react'
 
 import { useV2ConnectedWalletHasPermission } from 'hooks/v2/contractReader/V2ConnectedWalletHasPermission'
 import { V2OperatorPermission } from 'models/v2/permissions'
 import useProjectQueuedFundingCycle from 'hooks/v2/contractReader/ProjectQueuedFundingCycle'
 
 import { useIsUserAddress } from 'hooks/IsUserAddress'
-
-import { ProjectPage } from 'models/project-visibility'
 
 import { useRouter } from 'next/router'
 
@@ -21,11 +19,9 @@ import { RelaunchFundingCycleBanner } from './banners/RelaunchFundingCycleBanner
 import V2ProjectHeaderActions from './V2ProjectHeaderActions'
 
 import V2ProjectInfo from './V2ProjectInfo'
-import V2ProjectSettings from './V2ProjectSettingsPage/V2ProjectSettings'
-
-const defaultPage: ProjectPage = 'info'
 
 export default function V2Project() {
+  const router = useRouter()
   const {
     projectId,
     projectMetadata,
@@ -35,22 +31,6 @@ export default function V2Project() {
     projectOwnerAddress,
     handle,
   } = useContext(V2ProjectContext)
-
-  const [activePage, setActivePage] = useState<ProjectPage>(defaultPage)
-
-  const router = useRouter()
-  useEffect(() => {
-    setActivePage(() => {
-      switch (router.query.page) {
-        case 'info':
-          return 'info'
-        case 'settings':
-          return 'settings'
-        default:
-          return defaultPage
-      }
-    })
-  }, [router.query.page])
 
   const [handleModalVisible, setHandleModalVisible] = useState<boolean>()
 
@@ -92,13 +72,10 @@ export default function V2Project() {
             : undefined
         }
       />
-      {activePage === 'info' && (
-        <V2ProjectInfo
-          handleModalVisible={handleModalVisible}
-          setHandleModalVisible={setHandleModalVisible}
-        />
-      )}
-      {activePage === 'settings' && <V2ProjectSettings />}
+      <V2ProjectInfo
+        handleModalVisible={handleModalVisible}
+        setHandleModalVisible={setHandleModalVisible}
+      />
     </Space>
   )
 }
